+++ apps/admin-dashboard/app/routes/_dashboard.applications.$id.accept.tsx
@@ -10,9 +10,9 @@ import {
   useLoaderData,
 } from '@remix-run/react';
 
+import { acceptApplication, getApplication } from '@oyster/core/applications';
 import { Button, Form, Modal } from '@oyster/ui';
 
-import { acceptApplication, getApplication } from '@/admin-dashboard.server';
 import { Route } from '@/shared/constants';
 import {
   admin,
