+++ apps/admin-dashboard/app/routes/_dashboard.applications.$id.email.tsx
@@ -11,6 +11,10 @@ import {
 } from '@remix-run/react';
 import { type z } from 'zod';
 
+import {
+  getApplication,
+  updateEmailApplication,
+} from '@oyster/core/applications';
 import { Application } from '@oyster/types';
 import {
   Button,
@@ -21,10 +25,6 @@ import {
   validateForm,
 } from '@oyster/ui';
 
-import {
-  getApplication,
-  updateEmailApplication,
-} from '@/admin-dashboard.server';
 import { Route } from '@/shared/constants';
 import {
   commitSession,
